requirements.txt
discord.py
phonenumbers
pytesseract
ipwhois
beautifulsoup4
configparser
langdetect
PyQt5
PyQtWebEngine
